Now 6bits per field in Custom Telemetry View

+++ src/open9xeeprom.h
@@ -576,7 +576,7 @@ PACK(typedef struct t_Open9xModelData_v205 {
   uint8_t   thrTraceSrc;
   uint8_t   modelId;
   uint8_t   frskyLines[4];
-  uint8_t   frskyLinesXtra;
+  uint16_t  frskyLinesXtra;
 
   operator ModelData();
   t_Open9xModelData_v205() { memset(this, 0, sizeof(t_Open9xModelData_v205)); }
